feat(console-app): add thread participants
